refactor: improved file naming schema
