+++ components/editor/SymbolSelect.vue
