+++ components/parser/Display.vue
@@ -19,20 +19,20 @@ withDefaults(defineProps<Props>(), {
 <template>
 	<div class="resultZone">
 		<div v-if="companions.length > 0">
-			<ParserPlayerDisplayHeading :player-type="headingName" />
+			<ParserDisplayHeading :player-type="headingName" />
 			<div class="companions">
 				<div v-for="i in [0, 1]" :key="i" class="companion--wrapper">
-					<ParserPlayerDisplayItem
+					<ParserDisplayItem
 						v-for="player in companions.slice([0, slicePosition][i], [slicePosition, companions.length][i])" :key="player.n"
 						:player-link="player.l" :player-name="player.n" :player-symbol="player.s" />
 				</div>
 			</div>
 		</div>
 		<div v-if="pastCompanions.length > 0">
-			<ParserPlayerDisplayHeading :player-type="headingName" :use-past-cmatw="true" />
+			<ParserDisplayHeading :player-type="headingName" :use-past-cmatw="true" />
 			<div class="companions">
 				<div v-for="i in [0, 1]" :key="i" class="companion--wrapper">
-					<ParserPlayerDisplayItem
+					<ParserDisplayItem
 						v-for="player in pastCompanions.slice([0, slicePosition][i], [slicePosition, pastCompanions.length][i])" :key="player.n"
 						:player-link="player.l" :player-name="player.n" :player-symbol="player.s" />
 				</div>
