+++ components/stats/GeneralDisplay.vue
