+++ components/stats/MuralDisplay.vue
