+++ components/stats/SymbolDisplay.vue
